requirements.txt
requests
selenium
python-slugify
argparse